fix: code review improvements
- Fix FTS index rebuild after reindexNote()
- Fix hybrid search key collision (use id instead of title)
- Add HTML escaping for table cell values
- Use NoteNotFoundError in reindexNote()
- Add tests for ID-based lookup, rebuildFtsIndex, TableOutOfBoundsError
- Refactor: split read.ts into read.ts, resolve.ts, conversion.ts
