+++ src/notes/read.ts
@@ -6,39 +6,14 @@
  */
 
 import { runJxa } from "run-jxa";
-import TurndownService from "turndown";
 import { createDebugLogger } from "../utils/debug.js";
-import type { NoteSuggestion } from "../errors/index.js";
+import { htmlToMarkdown } from "./conversion.js";
+
+// Re-export for backwards compatibility
+export { resolveNoteTitle, type ResolvedNote } from "./resolve.js";
 
 // Debug logging
 const debug = createDebugLogger("NOTES");
-
-// Initialize Turndown for HTML to Markdown conversion
-const turndownService = new TurndownService({
-  headingStyle: "atx",
-  codeBlockStyle: "fenced",
-  bulletListMarker: "-",
-});
-
-// Add custom rule to handle Apple Notes attachment placeholders
-turndownService.addRule("attachments", {
-  filter: (node) => {
-    // Apple Notes uses object tags for attachments
-    return node.nodeName === "OBJECT" || node.nodeName === "IMG";
-  },
-  replacement: (_content, node) => {
-    // Turndown uses its own Node type, cast to access attributes
-    const element = node as unknown as {
-      getAttribute: (name: string) => string | null;
-    };
-    const filename =
-      element.getAttribute("data-filename") ||
-      element.getAttribute("alt") ||
-      element.getAttribute("src")?.split("/").pop() ||
-      "unknown";
-    return `[Attachment: ${filename}]`;
-  },
-});
 
 // -----------------------------------------------------------------------------
 // Types
@@ -64,21 +39,6 @@ export interface NoteDetails extends NoteInfo {
   id: string;
 }
 
-export interface ResolvedNote {
-  /** Whether resolution was successful */
-  success: boolean;
-  /** The matched note (if exactly one match) */
-  note?: {
-    title: string;
-    folder: string;
-    id: string;
-  };
-  /** Error message if resolution failed */
-  error?: string;
-  /** Rich suggestions when multiple matches found (includes ID and created date) */
-  suggestions?: NoteSuggestion[];
-}
-
 // -----------------------------------------------------------------------------
 // Internal JXA helpers
 // -----------------------------------------------------------------------------
@@ -96,36 +56,6 @@ async function executeJxa<T>(code: string): Promise<T> {
       `JXA execution failed: ${error instanceof Error ? error.message : String(error)}`
     );
   }
-}
-
-/**
- * Convert HTML content to Markdown, handling Apple Notes specifics
- */
-function htmlToMarkdown(html: string): string {
-  if (!html) return "";
-
-  // Pre-process: handle Apple Notes specific markup
-  let processed = html;
-
-  // Replace attachment objects with placeholder text before Turndown
-  processed = processed.replace(
-    /<object[^>]*data-filename="([^"]*)"[^>]*>.*?<\/object>/gi,
-    "[Attachment: $1]"
-  );
-
-  // Handle inline images
-  processed = processed.replace(
-    /<img[^>]*(?:alt="([^"]*)")?[^>]*>/gi,
-    (_match, alt) => {
-      const filename = alt || "image";
-      return `[Attachment: ${filename}]`;
-    }
-  );
-
-  // Convert to Markdown
-  const markdown = turndownService.turndown(processed);
-
-  return markdown.trim();
 }
 
 // -----------------------------------------------------------------------------
@@ -179,7 +109,7 @@ export async function getAllNotes(): Promise<NoteInfo[]> {
 /**
  * Get a note by its title, with full content
  *
- * @param title - The note title (can be "folder/title" format for disambiguation)
+ * @param title - The note title (can be "folder/title" or "id:xxx" format)
  * @returns Note details with content, or null if not found
  */
 export async function getNoteByTitle(
@@ -488,133 +418,4 @@ export async function getAllFolders(): Promise<string[]> {
 
   debug(`Found ${folders.length} folders`);
   return folders;
-}
-
-/**
- * Resolve a note title input to a unique note
- *
- * Handles:
- * - "id:xxx" format for direct ID lookup
- * - Exact title match
- * - "folder/title" format for disambiguation
- * - Returns suggestions when multiple matches exist
- *
- * @param input - The note title, "folder/title", or "id:xxx" string
- * @returns Resolution result with success status, note info, or suggestions
- */
-export async function resolveNoteTitle(input: string): Promise<ResolvedNote> {
-  debug(`Resolving note title: ${input}`);
-
-  // Check for id:xxx format for direct ID lookup
-  if (input.startsWith("id:")) {
-    const noteId = input.slice(3);
-    debug(`ID prefix detected, looking up note by ID: ${noteId}`);
-    const noteDetails = await getNoteById(noteId);
-    if (!noteDetails) {
-      return {
-        success: false,
-        error: `Note not found with ID: "${noteId}"`,
-      };
-    }
-    return {
-      success: true,
-      note: {
-        id: noteDetails.id,
-        title: noteDetails.title,
-        folder: noteDetails.folder,
-      },
-    };
-  }
-
-  // Check for folder/title format
-  let targetFolder: string | null = null;
-  let targetTitle = input;
-
-  if (input.includes("/")) {
-    const parts = input.split("/");
-    targetFolder = parts.slice(0, -1).join("/");
-    targetTitle = parts[parts.length - 1];
-    debug(`Parsed folder: ${targetFolder}, title: ${targetTitle}`);
-  }
-
-  const escapedTitle = JSON.stringify(targetTitle);
-  const escapedFolder = targetFolder ? JSON.stringify(targetFolder) : "null";
-
-  const jxaCode = `
-    const app = Application('Notes');
-    app.includeStandardAdditions = true;
-
-    const targetTitle = ${escapedTitle};
-    const targetFolder = ${escapedFolder};
-
-    let foundNotes = [];
-    const folders = app.folders();
-
-    for (const folder of folders) {
-      const folderName = folder.name();
-
-      // Skip if folder filter is specified and doesn't match
-      if (targetFolder !== null && folderName !== targetFolder) {
-        continue;
-      }
-
-      const notes = folder.notes.whose({ name: targetTitle });
-
-      for (let i = 0; i < notes.length; i++) {
-        try {
-          const note = notes[i];
-          const props = note.properties();
-          foundNotes.push({
-            id: note.id(),
-            title: note.name(),
-            folder: folderName,
-            created: props.creationDate ? props.creationDate.toISOString() : ''
-          });
-        } catch (e) {
-          // Skip notes that can't be accessed
-        }
-      }
-    }
-
-    return JSON.stringify(foundNotes);
-  `;
-
-  const result = await executeJxa<string>(jxaCode);
-  const notes = JSON.parse(result) as Array<{
-    id: string;
-    title: string;
-    folder: string;
-    created: string;
-  }>;
-
-  if (notes.length === 0) {
-    debug("No matching notes found");
-    return {
-      success: false,
-      error: `Note not found: "${input}"`,
-    };
-  }
-
-  if (notes.length === 1) {
-    debug(`Resolved to unique note: ${notes[0].folder}/${notes[0].title}`);
-    return {
-      success: true,
-      note: notes[0],
-    };
-  }
-
-  // Multiple matches - return rich suggestions with ID and created date
-  const suggestions: NoteSuggestion[] = notes.map((n) => ({
-    id: n.id,
-    folder: n.folder,
-    title: n.title,
-    created: n.created,
-  }));
-  debug(`Multiple matches found: ${suggestions.map(s => `${s.folder}/${s.title}`).join(", ")}`);
-
-  return {
-    success: false,
-    error: `Multiple notes found with title "${targetTitle}". Use ID prefix to specify.`,
-    suggestions,
-  };
 }
